Print status as we send files TO the device
Normally the status is only printed as we receive files from the device - but on very large backups (e.g. with a 256MB manifest) that can make it appear that the backup has stalled.
And, if slow enough (more than 15m), can cause us to think the backup has locked up - forcing us to terminate it early.